refactor(core/rust): move iter_into_array to micropython::utils
