+++ core/embed/rust/src/micropython/ffi.rs
@@ -2,5 +2,6 @@
 #![allow(non_upper_case_globals)]
 #![allow(dead_code)]
 #![allow(clippy::unnecessary_cast)]
+#![allow(unsafe_op_in_unsafe_fn)]
 
 include!(concat!(env!("OUT_DIR"), "/micropython.rs"));
